Added new 'DependencyInjectionContainerToolkitPackage' and updated IToolkitServiceProvider and implementation to be generic.
